+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Bug fixes:
 
 * `C_numeric_deriv` gives wrong results of gradient #54
 * `tcrossprod` with a single vector #56
+* whenever possible, errors are propagated to the FastR embedder
 
 # 1.0 RC 13